+++ index.html
@@ -7,7 +7,6 @@
   <link rel="stylesheet" href="css/style.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
   <title>Кекстаграм</title>
-  <script src="functions.js"></script>
 </head>
 
 <body>
@@ -235,6 +234,6 @@
       </section>
     </template>
 
-    <script src="./js/main.js"></script>
+    <script src="./js/main.js" type="module" ></script>
 </body>
 </html>
