+++ js/const.js
@@ -1,0 +1,42 @@
+export const DESCRIPTIONS = [
+  'Утро',
+  'Машина',
+  'Природа',
+  'Котик'
+];
+
+export const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+export const NAMES = [
+  'Юрий',
+  'Артем',
+  'Иван',
+  'Мария',
+];
+
+export const LIKES = {
+  MIN: 15,
+  MAX: 200
+};
+
+export const COMMENTS_COUNT = {
+  MIN: 0,
+  MAX: 30
+};
+
+export const COMMENTS_ID = {
+  MIN: 10000,
+  MAX: 100000
+};
+
+export const AVATAR = {
+  MIN: 1,
+  MAX: 6
+};
