+++ js/utils.js
@@ -1,0 +1,8 @@
+export const getRandomInteger = (a, b) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+export const getRandomElement = (items) => items[getRandomInteger(0, items.length - 1)];
